Avoid f-strings for now
This commit removes the use of f strings from tomlgen module. The package
metadata didn't set a minimum python version and fstrings don't work
with python 3.5. To enable having the last release of setuptools-rust
prior to setting the python_requires work with python 3.5 we need to
have a release which will work for 3.5. With this change setuptools-rust
should work fine with Python 3.5. So we can release this and then set
the minimum supported python version to 3.6 and reintroduce the use of
fstrings.

Fixes #85