Fix segfaults caused by modifying existing shared library
When the built shared library is copied to the correct location, it
actually *modifies* any existing shared library, which causes any
running process using it to segfault.

To fix this, we first delete any existing shared library (which is safe
to do) and then copy the newly built shared library to the correct
location as a new file.

Related to #257