Delay imports of non-stdlib dependencies until time of use
This is a bit of a hack, but we do this to ensure that when setuptools
loads entrypoint based hooks it cannot (will not?) crash.

The issue is that setuptools plugins are autoloaded, whether any given
project uses them at all or not. So if setuptools-rust is installed,
setuptools always tries to use it, and crashes if setuptools-rust is
broken.

Of course, setuptools-rust can't be broken, because it's a wonderful
project.

BUT.

As it happens, third-party vendors providing setuptools-rust can get
into a situation where multiple packages need to be installed, including
setuptools-rust, and also build yet other packages from source. In the
middle of this, setuptools-rust itself could be installed but in
"half-configured" state, i.e. its dependencies were queued for
afterwards due to complicated dependency graph magic.

In such a scenario, it should be nominally all right to have an inert
package installed, since if nothing actually uses setuptools-rust it
doesn't need to *work* yet. And in fact, it is all right, as long as
setuptools can import the autoloaded plugin hooks (and do nothing with
them).

Bug: https://bugs.gentoo.org/933553