Try extra hard to pick up an existing tomli library from setuptools
Since setuptools-rust already depends on setuptools, it is reasonable to
assume that even if tomli isn't installed, setuptools is. And setuptools
includes a vendored copy of tomli.

If the copy in setuptools has been devendored, it will be available via
"tomli". If it isn't devendored, it will be available via
"setuptools.extern.tomli" unless setuptools changes their vendoring
approach which has lasted many years so far. Either way, we are sure to
have a fallback tomli without explicitly depending on one, which means
one less dependency to install in the common case.